Reorganize e2e structure
